Refactor PostHog middleware and exception handling into dedicated module for cleaner integration.
