+++ backend/src/middleware/__init__.py
@@ -1,0 +1,1 @@
+from .posthog import posthog, posthog_middleware, http_exception_handler
